# docs/api/openapi.v0.1.yaml
openapi: 3.1.0
info:
  title: Knowledge Base Agent Assistant API
  version: 0.1.0
  description: v0.1 frontend-backend API contract for the internal Knowledge Base Agent Assistant.
servers:
  - url: http://localhost:8000/api/v1
    description: Local backend
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
...
  /auth/consumer-session:
    post:
      tags: [Auth]
      security: []
      summary: Create a passwordless consumer user session
      description: Creates a passwordless session for the selected active user-role account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerSessionRequest"
      responses:
        "200":
          description: Consumer session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Refreshed token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Error"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Error"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
...
  /knowledge-bases/{kb_id}/files:
    get:
      tags: [Files]
      summary: List files in knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/FileStatus"
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
  /knowledge-bases/{kb_id}/files/upload:
    post:
      tags: [Files]
      summary: Upload files to knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  maxItems: 50
                  items:
                    type: string
                    format: binary
                force:
                  type: boolean
                  default: false
      responses:
        "202":
          description: Files accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResponse"
        "409":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
  /files/{file_id}:
    get:
      tags: [Files]
      summary: Get file
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
    delete:
      tags: [Files]
      summary: Soft delete file
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: Deleted
  /files/{file_id}/status:
    get:
      tags: [Files]
      summary: Get file parse/index status
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: File status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileStatusResponse"
  /files/{file_id}/summary:
    get:
      tags: [Files]
      summary: Get current document summary
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Document summary and chunk extraction progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSummary"
  /files/{file_id}/summary/retry:
    post:
      tags: [Files]
      summary: Retry or rebuild current document summary
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
      responses:
        "202":
          description: Summary task queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSummary"
  /files/{file_id}/retry-parse:
    post:
      tags: [Files]
      summary: Retry file parse
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "202":
          description: Retry job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseJob"
  /files/{file_id}/chunks:
    get:
      tags: [Files]
      summary: List chunks for file
      parameters:
        - $ref: "#/components/parameters/FileId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Chunk list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    KnowledgeBaseId:
      name: kb_id
      in: path
      required: true
      schema:
        type: string
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ISODateTime:
      type: string
      format: date-time
    UserRole:
      type: string
      enum: [admin, user]
...
    FileStatus:
      type: string
      enum: [uploaded, queued, processing, indexed, partially_indexed, failed, deleting, deleted]
    ParseJobStatus:
      type: string
      enum: [queued, parsing, normalizing, chunking, embedding, indexing, indexed, partially_indexed, failed, cancelled]
...
    ErrorEnvelope:
      type: object
      required: [code, message, details, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorEnvelope"
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    ConsumerSessionRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
...
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 1800
    LoginResponse:
      allOf:
        - $ref: "#/components/schemas/TokenResponse"
        - type: object
          required: [user]
          properties:
            user:
              $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, username, display_name, role, is_active, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        is_active:
          type: boolean
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        last_login_at:
          anyOf:
            - $ref: "#/components/schemas/ISODateTime"
            - type: "null"
...
    PaginatedResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    File:
      type: object
      required: [id, knowledge_base_id, file_name, file_ext, size_bytes, file_hash, status, created_by, created_at, updated_at]
      properties:
        id:
          type: string
        knowledge_base_id:
          type: string
        file_name:
          type: string
        file_ext:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        file_hash:
          type: string
        status:
          $ref: "#/components/schemas/FileStatus"
        latest_parse_job_id:
          type:
            - string
            - "null"
        created_by:
          type: string
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        updated_at:
          $ref: "#/components/schemas/ISODateTime"
    ParseJob:
      type: object
      required: [id, file_id, status, progress, error_code, error_message, logs, created_at, updated_at]
      properties:
        id:
          type: string
        file_id:
          type: string
        status:
          $ref: "#/components/schemas/ParseJobStatus"
        progress:
          type: integer
          minimum: 0
          maximum: 100
        error_code:
          type:
            - string
            - "null"
        error_message:
          type:
            - string
            - "null"
        logs:
          type:
            - object
            - "null"
          additionalProperties: true
        started_at:
          anyOf:
            - $ref: "#/components/schemas/ISODateTime"
            - type: "null"
        finished_at:
          anyOf:
            - $ref: "#/components/schemas/ISODateTime"
            - type: "null"
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        updated_at:
          $ref: "#/components/schemas/ISODateTime"
    FileUploadItem:
      type: object
      required: [file, parse_job]
      properties:
        file:
          $ref: "#/components/schemas/File"
        parse_job:
          $ref: "#/components/schemas/ParseJob"
    FileUploadWarning:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    FileUploadResponse:
      type: object
      required: [uploaded, warnings]
      properties:
        uploaded:
          type: array
          items:
            $ref: "#/components/schemas/FileUploadItem"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/FileUploadWarning"
    FileStatusResponse:
      type: object
      required: [file_id, file_status, latest_parse_job]
      properties:
        file_id:
          type: string
        file_status:
          $ref: "#/components/schemas/FileStatus"
        latest_parse_job:
          anyOf:
            - $ref: "#/components/schemas/ParseJob"
            - type: "null"
    FileListResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/File"
    Chunk:
      type: object
      required: [id, file_id, knowledge_base_id, content, source_locator, token_count, is_active, created_at]
      properties:
        id:
          type: string
        file_id:
          type: string
        knowledge_base_id:
          type: string
        content:
          type: string
        source_locator:
          type: string
        token_count:
          type: integer
        is_active:
          type: boolean
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        knowledge_extraction:
          anyOf:
            - $ref: "#/components/schemas/ChunkKnowledgeExtraction"
            - type: "null"
    ChunkKnowledgeExtraction:
      type: object
      required: [status, prompt_version, attempt_count]
      properties:
        status:
          type: string
          enum: [pending, running, completed, failed]
        extraction:
          type:
            - object
            - "null"
          additionalProperties: true
        model_name:
          type:
            - string
            - "null"
        prompt_version:
          type: string
        attempt_count:
          type: integer
        error_code:
          type:
            - string
            - "null"
        error_message:
          type:
            - string
            - "null"
    DocumentSummary:
      type: object
      required:
        - file_id
        - status
        - chunk_total
        - chunk_completed
        - chunk_succeeded
        - chunk_failed
        - failed_chunk_ids
        - chunk_prompt_version
        - document_prompt_version
        - reduction_level
      properties:
        file_id:
          type: string
        parse_job_id:
          type:
            - string
            - "null"
        status:
          type: string
          enum: [pending, running, completed, partially_completed, failed, not_ready]
        summary:
          type:
            - string
            - "null"
        chunk_total:
          type: integer
        chunk_completed:
          type: integer
        chunk_succeeded:
          type: integer
        chunk_failed:
          type: integer
        failed_chunk_ids:
          type: array
          items:
            type: string
        model_name:
          type:
            - string
            - "null"
        chunk_prompt_version:
          type: string
        document_prompt_version:
          type: string
        reduction_level:
          type: integer
        error_code:
          type:
            - string
            - "null"
        error_message:
          type:
            - string
            - "null"
...
    ChunkListResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Chunk"
...